+++ project/src/pages/Room.tsx
@@ -7,6 +7,7 @@ import { ReviewList } from '../components/review-list';
 import { Map } from '../components/map';
 import { useAppSelector, useAppDispatch } from '../hooks';
 import { MAX_RATING_VALUE } from '../constants';
+import { Header } from '../components/header';
 import {
   fetchCommentsAction,
   fetchHotelAction,
@@ -41,40 +42,9 @@ export const Room = (): JSX.Element => {
 
   return (
     <div className='page'>
-      <header className='header' ref={headerRef}>
-        <div className='container'>
-          <div className='header__wrapper'>
-            <div className='header__left'>
-              <a className='header__logo-link' href='main.html'>
-                <img
-                  className='header__logo'
-                  src='img/logo.svg'
-                  alt='6 cities logo'
-                  width='81'
-                  height='41'
-                />
-              </a>
-            </div>
-            <nav className='header__nav'>
-              <ul className='header__nav-list'>
-                <li className='header__nav-item user'>
-                  <div className='header__nav-profile'>
-                    <div className='header__avatar-wrapper user__avatar-wrapper'></div>
-                    <span className='header__user-name user__name'>
-                      Oliver.conner@gmail.com
-                    </span>
-                  </div>
-                </li>
-                <li className='header__nav-item'>
-                  <a className='header__nav-link' href='http://localhost:3000'>
-                    <span className='header__signout'>Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <div ref={headerRef}>
+        <Header />
+      </div>
 
       {offer ? (
         <main className='page__main page__main--property'>
